consolidate tests into one file
